+++ tests/test_06_ingest.py
@@ -17,6 +17,13 @@ from conftest import (getConfig, gettestdata, icat_version,
 
 logger = logging.getLogger(__name__)
 
+# There seem to be a bug in older icat.server versions when searching
+# for a single boolean attribute in a query.
+skip_single_boolean = icat_version < "4.11.0"
+
+def print_xml(root):
+    print('\n', etree.tostring(root, pretty_print=True).decode(), sep='')
+
 def get_test_investigation(client):
     query = Query(client, "Investigation", conditions={
         "name": "= '12100409-ST'",
@@ -73,8 +80,24 @@ def schemadir(monkeypatch):
     monkeypatch.setattr(IngestReader, "SchemaDir", testdatadir)
 
 
+class EnvironmentIngestReader(IngestReader):
+    """Modified version of IngestReader
+    - Allow custom environment settings to be included.
+    - Capture the ingest data after injection of the environment in an
+      attribute.
+    """
+    _add_env = dict()
+    def get_environment(self, client):
+        env = super().get_environment(client)
+        env.update(self._add_env)
+        return env
+    def add_environment(self, client, ingest_data):
+        super().add_environment(client, ingest_data)
+        self._ingest_data = ingest_data
+
+
 class MyIngestReader(IngestReader):
-    """Testting a customized IngestReader
+    """Testing a customized IngestReader
     """
     XSD_Map = {
         ('icatingest', '1.0'): "ingest-10.xsd",
@@ -99,6 +122,11 @@ cases = [
         metadata = gettestdata("metadata-4.4-inl.xml"),
         checks = {
             "testingest_inl_1": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 2.7 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -115,6 +143,11 @@ cases = [
                  2.74103),
             ],
             "testingest_inl_2": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 5.1 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -138,6 +171,11 @@ cases = [
         metadata = gettestdata("metadata-5.0-inl.xml"),
         checks = {
             "testingest_inl5_1": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 2.7 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -162,6 +200,11 @@ cases = [
                  2.74103),
             ],
             "testingest_inl5_2": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 5.1 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -196,6 +239,11 @@ cases = [
         metadata = gettestdata("metadata-4.4-sep.xml"),
         checks = {
             "testingest_sep_1": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 2.7 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -212,6 +260,11 @@ cases = [
                  2.74103),
             ],
             "testingest_sep_2": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 5.1 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -235,6 +288,11 @@ cases = [
         metadata = gettestdata("metadata-5.0-sep.xml"),
         checks = {
             "testingest_sep5_1": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 2.7 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -259,6 +317,11 @@ cases = [
                  2.74103),
             ],
             "testingest_sep5_2": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "Dy01Cp02 at 5.1 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -294,6 +357,11 @@ cases = [
         metadata = gettestdata("metadata-sample.xml"),
         checks = {
             "testingest_sample_1": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "ab3465 at 2.7 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -308,6 +376,11 @@ cases = [
                  "ab3465"),
             ],
             "testingest_sample_2": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "ab3465 at 5.1 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -322,6 +395,11 @@ cases = [
                  "ab3465"),
             ],
             "testingest_sample_3": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "ab3466 at 2.7 K"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -336,6 +414,11 @@ cases = [
                  "ab3466"),
             ],
             "testingest_sample_4": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 False),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "raw"),
                 ("SELECT ds.description FROM Dataset ds WHERE ds.id = %d",
                  "reference"),
                 ("SELECT ds.startDate FROM Dataset ds WHERE ds.id = %d",
@@ -361,7 +444,9 @@ def test_ingest_schema(client, investigation, schemadir, case):
     datasets = []
     for name in case.data:
         datasets.append(client.new("Dataset", name=name))
-    reader = IngestReader(client, case.metadata, investigation)
+    reader = EnvironmentIngestReader(client, case.metadata, investigation)
+    print_xml(reader._ingest_data)
+    print_xml(reader.infile)
     with get_icatdata_schema().open("rb") as f:
         schema = etree.XMLSchema(etree.parse(f))
     schema.assertValid(reader.infile)
@@ -385,6 +470,8 @@ def test_ingest(client, investigation, samples, schemadir, case):
         })
         ds = client.assertedSearch(query)[0]
         for query, res in case.checks[name]:
+            if skip_single_boolean and isinstance(res, bool):
+                continue
             assert client.assertedSearch(query % ds.id)[0] == res
 
 io_metadata = NamedBytesIO("""<?xml version='1.0' encoding='UTF-8'?>
@@ -446,6 +533,8 @@ def test_ingest_fileobj(client, investigation, samples, schemadir, case):
         })
         ds = client.assertedSearch(query)[0]
         for query, res in case.checks[name]:
+            if skip_single_boolean and isinstance(res, bool):
+                continue
             assert client.assertedSearch(query % ds.id)[0] == res
 
 
@@ -637,6 +726,65 @@ def test_ingest_error_searcherr(client, investigation, schemadir, case):
     logger.info("Raised %s: %s", exc.type.__name__, exc.value)
 
 
+classattr_metadata = NamedBytesIO("""<?xml version='1.0' encoding='UTF-8'?>
+<icatingest version="1.1">
+  <head>
+    <date>2024-10-11T10:51:26+02:00</date>
+    <generator>metadata-writer 0.27a</generator>
+  </head>
+  <data>
+    <dataset id="Dataset_1">
+      <name>testingest_classattr_1</name>
+      <description>Auxiliary data</description>
+      <startDate>2022-02-03T15:40:12+01:00</startDate>
+      <endDate>2022-02-03T17:04:22+01:00</endDate>
+    </dataset>
+  </data>
+</icatingest>
+""".encode("utf8"), "classattr_metadata")
+classattr_cases = [
+    Case(
+        data = ["testingest_classattr_1"],
+        metadata = classattr_metadata,
+        checks = {
+            "testingest_classattr_1": [
+                ("SELECT ds.complete FROM Dataset ds WHERE ds.id = %d",
+                 True),
+                (("SELECT t.name FROM DatasetType t JOIN t.datasets AS ds "
+                  "WHERE ds.id = %d"),
+                 "other"),
+            ],
+        },
+        marks = (),
+    ),
+]
+@pytest.mark.parametrize("case", [
+    pytest.param(c, id=c.metadata.name, marks=c.marks) for c in classattr_cases
+])
+@pytest.mark.skipif(skip_single_boolean, reason="Bug in icat.server")
+def test_ingest_classattr(monkeypatch, client, investigation, schemadir, case):
+    """Test overriding prescribed values set in IngestReader class attributes.
+    """
+    monkeypatch.setattr(IngestReader, "Dataset_complete", "true")
+    monkeypatch.setattr(IngestReader, "DatasetType_name", "other")
+    datasets = []
+    for name in case.data:
+        datasets.append(client.new("Dataset", name=name))
+    reader = IngestReader(client, case.metadata, investigation)
+    reader.ingest(datasets, dry_run=True, update_ds=True)
+    for ds in datasets:
+        ds.create()
+    reader.ingest(datasets)
+    for name in case.checks.keys():
+        query = Query(client, "Dataset", conditions={
+            "name": "= '%s'" % name,
+            "investigation.id": "= %d" % investigation.id,
+        })
+        ds = client.assertedSearch(query)[0]
+        for query, res in case.checks[name]:
+            assert client.assertedSearch(query % ds.id)[0] == res
+
+
 customcases = [
     Case(
         data = ["testingest_custom_icatingest_1"],
@@ -710,6 +858,8 @@ def test_custom_ingest(client, investigation, samples, schemadir, case):
         })
         ds = client.assertedSearch(query)[0]
         for query, res in case.checks[name]:
+            if skip_single_boolean and isinstance(res, bool):
+                continue
             assert client.assertedSearch(query % ds.id)[0] == res
 
 
@@ -727,22 +877,30 @@ env_cases = [
 def test_ingest_env(monkeypatch, client, investigation, schemadir, case):
     """Test using the _environment element.
 
-    Applying a custom XSLT that extracts an attribute from the
-    _environment element that is injected by IngestReader into the
-    input data and puts that values into the head element of the
-    transformed input.  This is to test that adding the _environment
-    element works and it is in principle possible to make use of the
-    values in the XSLT.
+    Add a custom attribute to the _environment that is injected by
+    IngestReader into the input data.  Apply a custom XSLT that
+    extracts attributes from the _environment element and puts the
+    values into the head element of the transformed input.  This is to
+    test that adding the _environment element works and it is in
+    principle possible to make use of the values in the XSLT.
     """
-    monkeypatch.setattr(IngestReader,
+    generator = "test_ingest_env (python-icat %s)" % icat.__version__
+    monkeypatch.setattr(EnvironmentIngestReader,
+                        "_add_env", dict(generator=generator))
+    monkeypatch.setattr(EnvironmentIngestReader,
                         "XSLT_Map", dict(icatingest="ingest-env.xslt"))
     datasets = []
     for name in case.data:
         datasets.append(client.new("Dataset", name=name))
-    reader = IngestReader(client, case.metadata, investigation)
+    reader = EnvironmentIngestReader(client, case.metadata, investigation)
+    print_xml(reader._ingest_data)
+    print_xml(reader.infile)
     with get_icatdata_schema().open("rb") as f:
         schema = etree.XMLSchema(etree.parse(f))
     schema.assertValid(reader.infile)
     version_elem = reader.infile.xpath("/icatdata/head/apiversion")
     assert version_elem
     assert version_elem[0].text == str(client.apiversion)
+    generator_elem = reader.infile.xpath("/icatdata/head/generator")
+    assert generator_elem
+    assert generator_elem[0].text == generator